refactor(admin): FSD 기반 디렉터리 구조 마이그레이션
- components/, routes/, hooks/, lib/ 제거
- app/(providers), pages/, widgets/, shared/ 레이어로 재편
- components.json alias 경로 shared/ 기준으로 업데이트
- apps/admin/CLAUDE.md 구조 가이드 문서 추가
