+++ apps/admin/src/pages/applications/ApplicationsPage.tsx
@@ -1,7 +1,7 @@
-import { Button } from "@/components/ui/button"
+import { Button } from "@/shared/ui/button"
 
 /** 지원자 관리 페이지 */
-export function Applications() {
+export function ApplicationsPage() {
   return (
     <div className="flex min-h-svh p-6">
       <div className="flex max-w-md min-w-0 flex-col gap-4 text-sm leading-loose">
