+++ apps/admin/src/pages/blog-posts/BlogPostsPage.tsx
@@ -1,0 +1,4 @@
+/** 블로그 관리 페이지 */
+export default function BlogPostsPage() {
+  return <></>
+}
