+++ apps/admin/src/pages/index.tsx
@@ -1,22 +1,21 @@
 import { createBrowserRouter, RouterProvider } from "react-router"
-import Login from "./pages/Login"
-
-import Semesters from "./pages/Semesters"
-import Reminders from "./pages/Reminders"
-import Projects from "./pages/Projects"
-import BlogPosts from "./pages/BlogPosts"
-import { ErrorPage } from "./pages/Error"
-import { DefaultLayout } from "@/components/layout/DefaultLayout"
-import { Applications } from "./pages/Applications"
+import LoginPage from "./login/LoginPage"
+import SemestersPage from "./semesters/SemestersPage"
+import RemindersPage from "./reminders/RemindersPage"
+import ProjectsPage from "./projects/ProjectsPage"
+import BlogPostsPage from "./blog-posts/BlogPostsPage"
+import { ErrorPage } from "./error/ErrorPage"
+import { AdminLayout } from "@/widgets/admin-layout/AdminLayout"
+import { ApplicationsPage } from "./applications/ApplicationsPage"
 
 /** 라우터 설정 (리액트 라우터 Data Mode 기반) */
 const router = createBrowserRouter([
   {
-    element: <DefaultLayout />,
+    element: <AdminLayout />,
     children: [
       {
         path: "/applications",
-        element: <Applications />,
+        element: <ApplicationsPage />,
         errorElement: <ErrorPage />,
       },
       {
@@ -28,23 +27,22 @@ const router = createBrowserRouter([
           //     throw new Response("Failed to load data", { status: res.status })
           //   }
         },
-        element: <Semesters />,
+        element: <SemestersPage />,
         errorElement: <ErrorPage />,
       },
       {
         path: "/reminders",
-        element: <Reminders />,
+        element: <RemindersPage />,
         errorElement: <ErrorPage />,
       },
       {
         path: "/projects",
-        element: <Projects />,
+        element: <ProjectsPage />,
         errorElement: <ErrorPage />,
       },
-
       {
         path: "/blog-posts",
-        element: <BlogPosts />,
+        element: <BlogPostsPage />,
         errorElement: <ErrorPage />,
       },
     ],
@@ -52,7 +50,7 @@ const router = createBrowserRouter([
   {
     path: "/",
     // 해당 라우트에 매칭되는 UI 컴포넌트(페이지)
-    element: <Login />,
+    element: <LoginPage />,
     // element 내부적으로 Errorboundary가 잡지 않는,
     // 이벤트 핸들러의 의한 에러를 제외하고 라우터에서 잡히는 에러에 대한 UI
     errorElement: <ErrorPage />,
