+++ apps/admin/src/pages/projects/ProjectsPage.tsx
@@ -1,0 +1,4 @@
+/** 프로젝트 관리 페이지 */
+export default function ProjectsPage() {
+  return <></>
+}
