+++ apps/admin/src/pages/reminders/RemindersPage.tsx
@@ -1,0 +1,4 @@
+/** 사전 알림 신청 페이지 */
+export default function RemindersPage() {
+  return <></>
+}
