+++ apps/admin/src/pages/semesters/SemestersPage.tsx
@@ -1,4 +1,4 @@
 /** 기수 관리 페이지 */
-export default function Semesters() {
+export default function SemestersPage() {
   return <></>
 }
