--- apps/admin/src/routes/pages/BlogPosts.tsx
@@ -1,3 +1,0 @@
-export default function BlogPosts() {
-  return <></>
-}
