--- apps/admin/src/routes/pages/Projects.tsx
@@ -1,3 +1,0 @@
-export default function Projects() {
-  return <></>
-}
