--- apps/admin/src/routes/pages/Reminders.tsx
@@ -1,4 +1,0 @@
-/** 사전 알림 신청 관리 페이지 */
-export default function Reminders() {
-  return <></>
-}
