+++ apps/admin/src/shared/hooks/useIsMobile.ts
