+++ apps/admin/src/shared/lib/paths.ts
