+++ apps/admin/src/shared/ui/drawer.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { Drawer as DrawerPrimitive } from "vaul"
 
-import { cn } from "@/lib/utils/cn"
+import { cn } from "@/shared/lib/cn"
 
 function Drawer({
   ...props
