+++ apps/admin/src/shared/ui/tooltip.tsx
@@ -1,6 +1,6 @@
 import { Tooltip as TooltipPrimitive } from "@base-ui/react/tooltip"
 
-import { cn } from "@/lib/utils/cn"
+import { cn } from "@/shared/lib/cn"
 
 function TooltipProvider({
   delay = 0,
